Download .nc files with tag data from the FB Google Drive

Default is to download all Ziphius SMRT tag datasets (or you can specify the tag(s) for which you want to download data). Data are downloaded from the FB project shared Google Drive.

download_drive_nc(
  tag_id = zc_smrt_tag_list,
  path = getwd(),
  email,
  overwrite = TRUE
)

Arguments

tag_id

Character string or vector with tag IDs to download (without "-cal.nc" and without additional file path information). Default: all SMRT ziphius tags.

path

Directory (quoted string) where .nc files should be stored. Can be one string, or a list the same length as tag_ids. Default: current working directory. Note: use "/" and not "\" to avoid headaches. Final "/" not necessary.

email

Email address (for FB Google Drive authentication). You may also be asked to sign in or verify your Google identity as this function runs.

overwrite

Logical; whether to overwrite existing files. Default is TRUE.

Value

Returns a "dribble" with information about the files and their location on google drive. (The function also downloads the requested files, of course.)