extract RL data

This function processes the "raw" format of RL data, allowing users to obtain RLs of desired types and in selected frequency bands. If you need direct access to the (more complicated) raw data, see download_drive_rls.

extract_rls(
  rl_file,
  ping_log_file,
  email,
  save_output = TRUE,
  path = getwd(),
  out_file,
  signal = "MFAS",
  overwrite = TRUE
)

Arguments

rl_file
name (with path, if needed) of locally-stored .csv file with raw RL data. If not provided, the default is to use download_drive_rls to obtain it from the FB Google Drive.

ping_log_file
name (with path, if needed) of locally-stored .csv file with raw RL data. If not provided, the default is to use download_drive_rls to obtain it from the FB Google Drive.

email
Email address (required for FB Google Drive authentication; optional if rl_file is provided). You may also be asked to sign in or verify your Google identity as this function runs.

save_output
Logical; whether or not to save results in a .csv file. Default is TRUE.

path
Quoted string with the path to the directory where you want to save the output file. Defaults to the current working directory. Final "/" not needed. Use "/" rather than "\" to avoid possible headaches.

out_file
Name (quoted string) for output .csv file with results. Optional; default file name is constructed based on signal type and frequency band requested.

signal
Quoted string (or vector of them) indicating which signal types to return RLs for. Options are one or more of: 'MFAS', 'Echosounder', 'Explosive'. Default: 'MFAS'. Note: CAS is currently marked as MFAS type. You can isolate CAS pings and events by getting MFAS RLs and then keeping only pings with duration 20 seconds or longer.

overwrite
Whether or not to overwrite an existing output file. Logical. Default: TRUE.

Value

Returns a data.frame with RLs (one row per ping). If save_output is true, also saves a csv file in directory path with filename out_file.csv with the results.

Details

For MFAS: Events are marked type "MFA". This also includes CAS exposures (which, if we have separated them out before, it is by choosing MFA pings of duration more than 20 seconds). Measurements were made in ANSI-standard 1/3 octave bands centered from 1-40kHz. For MFA events, bands with center frequencies less than 9kHz are considered. For max RMS level, Units are: dB re 1 muPa. Measured in 200 msec windows; reported level is the highest in any one window. (Following conventions of SOCAL BRS, 3S projects.) For Echosounder events, selected bands are between 10-14kHz. For SPLs in individual 1/3 octave bands, the RL is "missing" (NA) if the SNR in that band was less than 6dB (and/or if the frequency band in question is no relevant for the selected signal type). For Explosions, all 1/3 octave bands below 5 kHz are included.

Examples

extract_rls(email = "sld33@calvin.edu") # (this only works if you know sld33's password...)

#> File downloaded:

#> * RLs_3obank.csv <id: 1qrT31r411MRuhyOd0Iv6Brsphoufzn11>

#> Saved locally as:

#> * C:/Users/Stacy DeRuiter/Dropbox/FBtagtools/docs/reference/RLs_3obank.csv

#> File downloaded:

#> * qPing_log_corr_times_master.csv <id: 19MyAcRfJeSgRkOwhC5EsJmFOjzwOQw0Q>

#> Saved locally as:

#> * C:/Users/Stacy DeRuiter/Dropbox/FBtagtools/docs/reference/qPing_log_corr_times_master.csv

#> Error in bind_rows(rl_file_dribble, ping_file_dribble): could not find function "bind_rows"