extract timing info for SMRT wav files

This function pulls data on start times (and other metadata) about acoustic wav files recorded by SMRT tags. It assumes that you already have downloaded the nc files for the deployments of interest to you; if not, see download_drive_nc

extract_wav_times(tag_id = zc_smrt_tag_list, nc_path = getwd())

Arguments

tag_id
Deployment IDs for the tag(s) for which you want to pull information. Defaults to the Ziphius deployments in zc_smrt_tag_list.

nc_path
Quoted string with the path to the directory where nc files are stored. Defaults to current working directory.

Value

Returns a data.frame with information about each wav file for each deployment of interest. If there are multiple rows per wav file, that is indicative of a (filled or un-filled) gap; the first row for a given deployment gives the file start time.

Examples

extract_rls(email = "sld33@calvin.edu") # (this only works if you know sld33's password...)

#> File downloaded:

#> * RLs_3obank.csv <id: 1qrT31r411MRuhyOd0Iv6Brsphoufzn11>

#> Saved locally as:

#> * C:/Users/Stacy DeRuiter/Dropbox/FBtagtools/docs/reference/RLs_3obank.csv

#> File downloaded:

#> * qPing_log_corr_times_master.csv <id: 19MyAcRfJeSgRkOwhC5EsJmFOjzwOQw0Q>

#> Saved locally as:

#> * C:/Users/Stacy DeRuiter/Dropbox/FBtagtools/docs/reference/qPing_log_corr_times_master.csv

#> Error in bind_rows(rl_file_dribble, ping_file_dribble): could not find function "bind_rows"